[Debugging] Fix problems regarding nav-bar
Problem 1. How to position two elements in nav-bar.
Used "align-items: center; justify-content: space-around;"

Problem 2. The flex-direction of ul is not changing.
Apply "display: flex;" to its parent element. (In this case, <nav> tag #nav-bar)

Problem 3. nav-bar is covering contents that are behind.
Top content can be shown by giving padding-top to body. More important issue was when we navigated to different places in the web page. We added <a> tags with ids that were used by divs.
By adding another tag, the flow of the content was not interrupted.